Fix Style/OneClassPerFile violations
I ended up disabling this cop for one occurrence in
`lib/mocha/parameter_matchers/instance_methods.rb` to avoid this YARD warning:

    [warn]: in YARD::Handlers::Ruby::MixinHandler: Undocumentable mixin:
            YARD::Handlers::NamespaceMissingError for class in file
            'lib/mocha/parameter_matchers/instance_methods.rb':25:

  	25: Object.include(Mocha::ParameterMatchers::InstanceMethods)

This wasn't a problem for the occurrences in `lib/mocha/inspect.rb`, because
this file is not included in `.yardopts`, i.e. YARD doesn't attempt to
generate documentation for these lines.

I think the warning is YARD saying that it can't generate documentation for
core Ruby classes like `Object`. A better solution would probably be to only
include the `Mocha::ParameterMatchers::InstanceMethods` module into `Object` at
Mocha "activation" time, i.e. at runtime. This would mean that YARD wouldn't
attempt to document the module's methods and the warning wouldn't arise.

Of course a better solution would be to avoid monkey-patching `Object`
altogether, but that's a whole other can of worms!
