+++ test/unit/mockery_test.rb
@@ -6,17 +6,6 @@ require 'mocha/state_machine'
 require 'mocha/expectation_error_factory'
 require 'simple_counter'
 
-class MockeryNeverSetupTest < Mocha::TestCase
-  include Mocha
-
-  def test_should_raise_not_initialized_error
-    Mockery.instance_variable_set(:@instances, nil)
-    assert_raises(NotInitializedError) do
-      Mockery.teardown
-    end
-  end
-end
-
 class MockeryTest < Mocha::TestCase
   include Mocha
